+++ matlab-kernel/src/main/java/com/tp/matlab/kernel/domain/result/FamResult.java
@@ -13,10 +13,10 @@ public class FamResult {
      * 频率
      */
     @NonNull
-    private final BigDecimal pl;
+    private final Double pl;
     /**
      * 幅值
      */
     @NonNull
-    private final BigDecimal fz;
+    private final Double fz;
 }
